# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: "MedRec Platform - Physician API"
host: localhost:10010
basePath: "/"
schemes:
- "http"
- "https"
consumes:
  - application/json
produces:
  - application/json
paths:
  /physicians/{id}:
    x-swagger-router-controller: physician
    get:
      summary: "Finds a Physician by Id"
      description: "Returns a single Physician"
      operationId: getOne
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetPhysicianResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      summary: "Updates an existing Physician"
      description: "Returns a Responseful execution message"
      operationId: update
      parameters:
        - name: id
          description: Physician id
          type: string
          in: path
          required: true
        - name: title
          description: Physician properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Physician"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"          
    delete:
      summary: "Deletes a Physician by Id"
      description: "Returns a Responseful execution message"
      operationId: delPhysician
      parameters:
        - name: id
          description: Physician id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
...
definitions:
  Physician:
    type: "object"
    required:
    - "FullName"
    - "MedicalSpeciality"
    properties:
      FullName:
        type: "string"
        example: "Mark Ridless"
      MedicalSpeciality:
        type: "string"
        example: "Geriatric"
  GetPhysicianResponse:
    required:
      - id
      - "FullName"
      - "MedicalSpeciality"
    properties:
      id:
        type: string
      FullName: 
        type: string
      MedicalSpeciality:
        type: number
...
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string